+++ core-ng/src/main/java/core/framework/internal/http/HTTPClientImpl.java
@@ -10,6 +10,7 @@ import core.framework.http.HTTPResponse;
 import core.framework.impl.log.filter.MapLogParam;
 import core.framework.log.ActionLogContext;
 import core.framework.log.Markers;
+import core.framework.util.ASCII;
 import core.framework.util.Maps;
 import core.framework.util.StopWatch;
 import core.framework.util.Strings;
@@ -17,6 +18,7 @@ import core.framework.util.Threads;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
@@ -29,9 +31,7 @@ import java.net.http.HttpTimeoutException;
 import java.time.Duration;
 import java.util.List;
 import java.util.Map;
-import java.util.TreeMap;
-
-import static java.lang.String.CASE_INSENSITIVE_ORDER;
+import java.util.zip.GZIPInputStream;
 
 /**
  * @author neo
@@ -111,20 +111,34 @@ public final class HTTPClientImpl implements HTTPClient {
     HTTPResponse response(HttpResponse<byte[]> httpResponse) {
         int statusCode = httpResponse.statusCode();
         logger.debug("[response] status={}", statusCode);
-        Map<String, String> headers = new TreeMap<>(CASE_INSENSITIVE_ORDER);
-        for (Map.Entry<String, List<String>> entry : httpResponse.headers().map().entrySet()) {
+        Map<String, List<String>> httpHeaders = httpResponse.headers().map();
+        Map<String, String> headers = Maps.newHashMapWithExpectedSize(httpHeaders.size());
+        for (Map.Entry<String, List<String>> entry : httpHeaders.entrySet()) {
             String name = entry.getKey();
             if (!Strings.startsWith(name, ':')) {   // not put pseudo headers
-                headers.put(name, entry.getValue().get(0));
+                headers.put(ASCII.toLowerCase(name), entry.getValue().get(0));
             }
         }
         logger.debug("[response] headers={}", new MapLogParam(headers));
 
-        byte[] body = httpResponse.body();
+        byte[] body = decodeBody(headers.get(HTTPHeaders.CONTENT_ENCODING), httpResponse.body());
+
         HTTPStatus status = parseHTTPStatus(statusCode);
         var response = new HTTPResponse(status, headers, body);
         logger.debug("[response] body={}", BodyLogParam.param(body, response.contentType));
         return response;
+    }
+
+    byte[] decodeBody(String encoding, byte[] body) {
+        // only support gzip, deflate is less popular
+        if ("gzip".equals(encoding)) {
+            try (var stream = new GZIPInputStream(new ByteArrayInputStream(body))) {
+                return stream.readAllBytes();
+            } catch (IOException e) {
+                throw new HTTPClientException("failed to decode body", "INVALID_BODY", e);
+            }
+        }
+        return body;
     }
 
     HttpRequest httpRequest(HTTPRequest request) {