+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Change log
 ### 6.9.4 (10/11/2018 - )
 * util: removed InputStreams, with JDK 11, inputStream added readAllBytes() method to read all bytes
+* httpclient: support gzip content-type
 
 ### 6.9.3 (10/8/2018 - 10/11/2018)
 * log: update slfj4 to 1.8, which switched to service provider model to bind logger factory and added java module support